[EA3-75] rename : dto 패키지 구조 변경 및 리팩터링
