+++ src/main/java/grep/neogul_coder/domain/calender/controller/PersonalCalenderController.java
@@ -1,12 +1,9 @@
 package grep.neogul_coder.domain.calender.controller;
 
-import grep.neogul_coder.domain.calender.controller.dto.PersonalCalenderRequest;
-import grep.neogul_coder.domain.calender.controller.dto.PersonalCalenderResponse;
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.tags.Tag;
+import grep.neogul_coder.domain.calender.controller.dto.requset.PersonalCalenderRequest;
+import grep.neogul_coder.domain.calender.controller.dto.response.PersonalCalenderResponse;
 import grep.neogul_coder.global.response.ApiResponse;
 
-import java.time.LocalDateTime;
 import java.util.List;
 import org.springframework.web.bind.annotation.*;
