+++ src/main/java/grep/neogul_coder/domain/calender/controller/dto/requset/TeamCalenderRequest.java
@@ -1,4 +1,4 @@
-package grep.neogul_coder.domain.calender.controller.dto;
+package grep.neogul_coder.domain.calender.controller.dto.requset;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.time.LocalDateTime;
