+++ src/main/java/grep/neogul_coder/domain/calender/controller/dto/response/TeamCalenderResponse.java
@@ -1,4 +1,4 @@
-package grep.neogul_coder.domain.calender.controller.dto;
+package grep.neogul_coder.domain.calender.controller.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.time.LocalDateTime;
